requirements.txt
# ============================================
# SISTEMA DE TRADING ALGORÍTMICO CON IA
# Dependencias principales
# ============================================

# Deep Learning & RL
torch>=2.0.0
tensorflow>=2.13.0
stable-baselines3>=2.1.0
gymnasium>=0.29.0

# Data Processing & Analysis
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Technical Analysis
ta-lib>=0.4.28
pandas-ta>=0.3.14b0

# APIs de Brokers
python-binance>=1.0.19
ccxt>=4.1.0
MetaTrader5>=5.0.45
ib_insync>=0.9.86
websocket-client>=1.6.0
requests>=2.31.0

# Base de Datos
psycopg2-binary>=2.9.9
redis>=5.0.0
sqlalchemy>=2.0.0

# Monitoring & Logging
prometheus-client>=0.18.0
python-telegram-bot>=20.6
loguru>=0.7.2

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
schedule>=1.2.0
pytz>=2023.3
tqdm>=4.66.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Performance
numba>=0.58.0
joblib>=1.3.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.5.0

# Desktop GUI
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
pyqtgraph>=0.13.3
qdarkstyle>=3.2.0

# Backtesting
backtrader>=1.9.78.123
vectorbt>=0.26.0
